refactor(edgar): move from built-in to runtime plugin
PR #1267 fixed runtime-plugin role gating, so the original reason
for shipping edgar as a built-in (avoid prompt pollution) no
longer applies. Move edgar to packages/plugins/edgar-plugin/ to
match the convention every other domain-specific plugin (spotify,
recipe-book, todo, debug, bookmarks) follows.

Trade-off accepted explicitly: the config path becomes
URL-encoded (~/mulmoclaude/config/plugins/%40mulmoclaude%2Fedgar-plugin/
config.json) because runtime.files.config keeps each scoped npm
name as a single safe directory segment. Existing users re-enter
their name + email at the new path; no migration shipped.

Layout:
- packages/plugins/edgar-plugin/src/{index,definition,args,edgar,
  config}.ts — definePlugin factory, Zod schema (extracted into
  args.ts so tests don't pull in the whole plugin), client with
  serialised throttle + AbortController fetch, self-healing
  config helper.
- packages/plugins/edgar-plugin/test/{args_validation,throttle,
  config}.ts — 45 in-package tests; pin the regex security
  guards, throttle invariant, and missing-config payload shape.
- src/config/toolNames.ts — `edgar` added to HOST_TOOL_NAMES so
  TOOL_NAMES.edgar stays type-safe in roles.ts (mirrors the
  pattern for other preset runtime plugins).
- server/plugins/preset-list.ts — `@mulmoclaude/edgar-plugin`
  appended.

Removed: src/plugins/edgar/, server/edgar/, server/api/routes/
edgar.ts, the route mount in server/index.ts, and the host-side
test/edgar/ tests (replaced by in-package tests above). yarn
plugins:codegen drops edgar from the built-in barrels.

Checks: typecheck ✓, lint ✓, 4490 host tests ✓, 45 in-package
tests ✓, build ✓, format ✓.
