+++ packages/plugins/edgar-plugin/package.json
@@ -1,0 +1,34 @@
+{
+  "name": "@mulmoclaude/edgar-plugin",
+  "version": "0.1.0",
+  "description": "SEC EDGAR runtime plugin for MulmoClaude — wraps the public EDGAR API as one tool with kind-discriminated dispatch. Server-only; no Vue View.",
+  "type": "module",
+  "main": "./dist/index.js",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.js"
+    }
+  },
+  "files": [
+    "dist"
+  ],
+  "scripts": {
+    "build": "vite build",
+    "typecheck": "tsc --noEmit",
+    "test": "tsx --test test/test_*.ts"
+  },
+  "peerDependencies": {
+    "gui-chat-protocol": "^0.3.0",
+    "zod": "^4.3.6"
+  },
+  "devDependencies": {
+    "@types/node": "^22.10.0",
+    "tsx": "^4.19.2",
+    "typescript": "^6.0.3",
+    "vite": "^8.0.11",
+    "vite-plugin-dts": "^5.0.0",
+    "zod": "^4.4.3"
+  },
+  "license": "MIT"
+}
