+++ server/index.ts
@@ -15,7 +15,6 @@ import imageRoutes from "./api/routes/image.js";
 import attachmentRoutes from "./api/routes/attachment.js";
 import presentHtmlRoutes from "./api/routes/presentHtml.js";
 import chartRoutes from "./api/routes/chart.js";
-import edgarRoutes from "./api/routes/edgar.js";
 import rolesRoutes from "./api/routes/roles.js";
 import { DEFAULT_ROLE_ID } from "../src/config/roles.js";
 import mulmoScriptRoutes from "./api/routes/mulmo-script.js";
@@ -464,7 +463,6 @@ app.use(imageRoutes);
 app.use(attachmentRoutes);
 app.use(presentHtmlRoutes);
 app.use(chartRoutes);
-app.use(edgarRoutes);
 app.use(rolesRoutes);
 app.use(mulmoScriptRoutes);
 app.use(wikiRoutes);
